pyproject.toml
[project]
name = "energy-forecasting"
version = "0.1.0"
description = "Energy Price Forecasting with ML Model Benchmarking"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "keras>=3.7.0",
    "matplotlib>=3.9.0",
    "mlflow>=2.18.0",
    "numpy>=1.26.4",
    "pandas>=2.2.3",
    "plotly>=5.24.0",
    "pylint>=3.3.2",
    "pytest>=8.3.4",
    "ruff>=0.11.0",
    "scikit-learn>=1.6.1,<1.7.0",
    "streamlit>=1.40.0",
    "tensorflow>=2.18.0",
    "xgboost>=2.1.0",
    "pyyaml>=6.0.2",
    "joblib>=1.4.0",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"
extend-include = ["*.ipynb"]
extend-exclude = ["justfile"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["E", "I", "F", "W"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"**/tests/*.py" = ["D103"]

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--maxfail=2 -rf -W ignore::DeprecationWarning"
